Ensure has_one associations cannot be orphaned
By default, if the has_one already exists and attributes are passed
without an id, a new record would be created leaving the old record
orphaned. This instead raises a helpful error message.

This is specific to has_one, not belongs_to, as belongs_to parent
records can still be valid even when the child moves parents.